+++ crates/sgf-gamedata/src/lib.rs
@@ -33,6 +33,7 @@ pub use registries::country_types::CountryTypes;
 pub use registries::defines::BorderDefines;
 pub use registries::deposits::Deposits;
 pub use registries::galaxy_shapes::GalaxyShapes;
+pub use registries::galaxy_sizes::GalaxySizes;
 pub use registries::gfx::Sprites;
 pub use registries::planet_classes::PlanetClasses;
 pub use registries::registry::Registry;
@@ -63,6 +64,7 @@ pub struct GameData {
     pub starbase_levels: Arc<StarbaseLevels>,
     pub ship_sizes: Arc<ShipSizes>,
     pub galaxy_shapes: Arc<GalaxyShapes>,
+    pub galaxy_sizes: Arc<GalaxySizes>,
     pub border: Arc<BorderDefines>,
     pub loc: Arc<Localisation>,
     pub diagnostics: Vec<Diagnostic>,
@@ -186,6 +188,7 @@ pub fn load(opts: &LoadOptions, progress: &mut dyn FnMut(Phase)) -> Result<GameD
     let ship_sizes = registry::load(&layout, &mut diagnostics);
     let starbase_levels = starbase_levels::load(&layout, &ship_sizes, &mut diagnostics);
     let galaxy_shapes = GalaxyShapes::load(&layout, &mut diagnostics);
+    let galaxy_sizes = GalaxySizes::load(&layout, &mut diagnostics);
     let sprites = gfx::load(&layout, &mut diagnostics);
     let colors = colors::load(&layout, &mut diagnostics);
     let border = BorderDefines::load(&layout, &mut diagnostics);
@@ -209,6 +212,7 @@ pub fn load(opts: &LoadOptions, progress: &mut dyn FnMut(Phase)) -> Result<GameD
         starbase_levels: Arc::new(starbase_levels),
         ship_sizes: Arc::new(ship_sizes),
         galaxy_shapes: Arc::new(galaxy_shapes),
+        galaxy_sizes: Arc::new(galaxy_sizes),
         border: Arc::new(border),
         loc: Arc::new(loc),
         diagnostics,